extsock: Use new HAVE_CJSON_FOR_EXTSOCK define
The extsock plugin's C code (extsock_client.c and extsock_plugin.c)
is updated to use the HAVE_CJSON_FOR_EXTSOCK preprocessor macro.

This macro was recently added to configure.ac to indicate if the
cJSON library is available. By using this define, the plugin can
more reliably conditionally compile cJSON-dependent features.

This change ensures that cJSON functionality is only compiled and
used when the library is confirmed to be present by the build
system, improving robustness.